refactor: move doublylink to new package
