+++ txpool/pool.go
@@ -11,6 +11,7 @@ import (
 	"github.com/pactus-project/pactus/types/tx"
 	"github.com/pactus-project/pactus/types/tx/payload"
 	"github.com/pactus-project/pactus/util/linkedmap"
+	"github.com/pactus-project/pactus/util/doublylink"
 	"github.com/pactus-project/pactus/util/logger"
 )
 
@@ -52,7 +53,7 @@ func (p *txPool) SetNewSandboxAndRecheck(sb sandbox.Sandbox) {
 	p.sandbox = sb
 	p.logger.Debug("set new sandbox")
 
-	var next *linkedmap.LinkNode[linkedmap.Pair[tx.ID, *tx.Tx]]
+	var next *doublylink.LinkNode[linkedmap.Pair[tx.ID, *tx.Tx]]
 	for _, pool := range p.pools {
 		for e := pool.HeadNode(); e != nil; e = next {
 			next = e.Next
