+++ util/doublylink/doublylink_test.go
@@ -1,4 +1,4 @@
-package linkedmap
+package doublylink
 
 import (
 	"testing"
